+++ src/presentation/controllers/signup.ts
@@ -2,8 +2,7 @@ import { badRequest, serverError } from '../helpers/http-helper';
 import { Controller } from '../protocols/controller';
 import { httpResponse, httpRequest } from '../protocols/http';
 import { EmailValidator } from '../protocols/email-validator';
-import MissingParamError from '../errors/missing-param-error';
-import InvalidParamError from '../errors/invalid-param-error';
+import { InvalidParamError, MissingParamError } from '../errors';
 
 export default class SignUpController implements Controller {
   private readonly emailValidator: EmailValidator;
