+++ src/presentation/errors/invalid-param-error.ts
@@ -1,4 +1,4 @@
-export default class InvalidParamError extends Error {
+export class InvalidParamError extends Error {
   constructor(readonly param: string) {
     super(`Invalid param: ${param}`);
     this.name = 'InvalidParamError';
