+++ src/presentation/errors/missing-param-error.ts
@@ -1,4 +1,4 @@
-export default class MissingParamError extends Error {
+export class MissingParamError extends Error {
   constructor(readonly param: string) {
     super(`Missing param: ${param}`);
     this.name = 'MissingParamError';
