+++ src/presentation/errors/server-error.ts
@@ -1,4 +1,4 @@
-export default class ServerError extends Error {
+export class ServerError extends Error {
   constructor() {
     super('Internal server error');
     this.name = 'ServerError';
